+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 # RxDB Changelog
 
 <!-- CHANGELOG NEWEST -->
+- IMPROVE performance of `.appendToArray()`
 - ADD iframe tracking to dev-mode plugin.
 <!-- ADD new changes here! -->